requirements.txt
# AFlow + ROLL Integration - Requirements
# Python 3.10+
# 基于当前可运行环境更新 (2025-11-24)

# Core Deep Learning (已验证兼容)
torch==2.6.0
transformers==4.57.1
peft==0.12.0

# Training & Optimization
numpy==2.2.6
pytorch-lightning==1.9.5

# Experiment Tracking
wandb==0.17.4

# Configuration & Data
pyyaml==6.0.1
tqdm==4.67.1

# Async & HTTP (AFlow需要)
aiohttp==3.13.2
openai==1.109.1
tenacity==8.5.0

# AFlow相关包
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0

# 性能优化
accelerate>=0.20.0
bitsandbytes>=0.41.0

# 开发 & 测试工具
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
